Refactor initial amount pages for Borrow/Deposit to remove duplication. Start refactor of confirmatin pages.
