+++ src/views/Deposit/DepositConfirm.tsx
@@ -3,17 +3,19 @@ import styled from "styled-components";
 import { useHistory, useRouteMatch } from "react-router-dom";
 import Page from "../../components/Page";
 import Button from "../../components/Button";
-import { marketData, IMarketData } from "../../utils/constants";
-import DepositOverview from "./DepositOverview";
 import {
   AgaveLendingABI__factory,
   Erc20abi__factory,
 } from "../../contracts";
+import { useAsset } from "../../hooks/asset";
+import { useBalance } from "../../hooks/balance";
+import { useApproved } from "../../hooks/approved";
 import { internalAddresses } from "../../utils/contracts/contractAddresses/internalAddresses";
-import { BigNumber } from "@ethersproject/bignumber";
-import { useQuery, useMutation, useQueryClient } from "react-query";
-import { useAppWeb3 } from "../../hooks/appWeb3";
+import { useMutation, useQueryClient } from "react-query";
 import { ethers } from "ethers";
+import { BigNumber } from "@ethersproject/bignumber";
+
+import DepositOverview from "./DepositOverview";
 
 const DepositConfirmWrapper = styled.div`
   height: 100%;
@@ -188,85 +190,35 @@ const DepositConfirm: React.FC = () => {
     assetName?: string | undefined;
     amount?: string | undefined;
   }>();
+
   const assetName = match.params.assetName;
-  const { account: address, library } = useAppWeb3();
-  const [wholeTokenAmount, setWholeTokenAmount] = useState<number>(0);
+  const [amount, setAmount] = useState<number>(0);
   // TODO: change this 'step' system to nested routes
   const [step, setStep] = useState(1);
 
-  const assetQueryKey = [assetName] as const;
-  const {
-    data: asset,
-  } = useQuery(
-    assetQueryKey,
-    async (ctx): Promise<IMarketData | undefined> => {
-      const [assetName]: typeof assetQueryKey = ctx.queryKey;
-      if (!assetName) {
-        return undefined;
-      }
+  const { asset, assetQueryKey } = useAsset(assetName);
+  const { balanceQueryKey } = useBalance(asset);
 
-      const asset = marketData.find((a) => a.name === match.params.assetName);
-      if (!asset) {
-        console.warn(`Asset ${match.params.assetName} not found`);
-        return;
+  useEffect(() => {
+    if (match.params && match.params.amount) {
+      try {
+        const parsed = Number(String(match.params.amount));
+        if (amount !== parsed) {
+          setAmount(parsed);
+        }
+      } catch {
+        // Don't set the number if the match path isn't one
       }
-
-      return asset;
-    },
-    {
-      initialData: undefined,
     }
-  );
+  }, [match, amount, setAmount]);
 
-  const balanceQueryKey = [address, library, asset] as const;
-  const {
-    data: balance,
-  } = useQuery(
-    balanceQueryKey,
-    async (ctx) => {
-      const [address, library, asset]: typeof balanceQueryKey = ctx.queryKey;
-      if (!address || !library || !asset) {
-        return undefined;
-      }
-      const contract = Erc20abi__factory.connect(
-        asset.contractAddress,
-        library.getSigner()
-      );
-      const tokenBalance = await contract.balanceOf(address);
-      return tokenBalance;
-    },
-    {
-      initialData: undefined,
-    }
-  );
+  const { approved: approval, approvedQueryKey } = useApproved(asset);
 
-  const approvedQueryKey = [address, library, asset] as const;
-  const {
-    data: approval,
-  } = useQuery(
-    approvedQueryKey,
-    async (ctx): Promise<BigNumber | undefined> => {
-      const [address, library, asset]: typeof approvedQueryKey = ctx.queryKey;
-      if (!address || !library || !asset) {
-        return undefined;
-      }
-      const contract = Erc20abi__factory.connect(
-        asset.contractAddress,
-        library.getSigner()
-      );
-      const allowance = await contract.allowance(address, internalAddresses.Lending);
-      return allowance;
-    },
-    {
-      initialData: BigNumber.from(0),
-    }
-  );
-
-  const approvalMutationKey = [...approvedQueryKey, wholeTokenAmount] as const;
+  const approvalMutationKey = [...approvedQueryKey, amount] as const;
   const approvalMutation = useMutation(
     approvalMutationKey,
     async (newValue) => {
-      const [address, library, asset, wholeTokenAmount] = approvalMutationKey;
+      const [address, library, asset, amount] = approvalMutationKey;
       if (!address || !library || !asset) {
         throw new Error("Account or asset details are not available");
       }
@@ -274,7 +226,7 @@ const DepositConfirm: React.FC = () => {
         asset.contractAddress,
         library.getSigner()
       );
-      const unitAmount = ethers.utils.parseEther(wholeTokenAmount.toString());
+      const unitAmount = ethers.utils.parseEther(amount.toString());
       const tx = await contract.approve(internalAddresses.Lending, unitAmount);
       const receipt = await tx.wait();
       return BigNumber.from(receipt.status ? unitAmount : 0);
@@ -284,7 +236,7 @@ const DepositConfirm: React.FC = () => {
         console.log("approvalMutation:onSuccess");
         await Promise.allSettled([
           // queryClient.invalidateQueries(approvedQueryKey), // Request that the approval query refreshes
-          queryClient.setQueryData(approvedQueryKey, ethers.utils.parseEther(wholeTokenAmount.toString())), // Update the approved amount query immediately
+          queryClient.setQueryData(approvedQueryKey, ethers.utils.parseEther(amount.toString())), // Update the approved amount query immediately
           queryClient.invalidateQueries(approvalMutationKey),
         ]);
         setStep(2);
@@ -292,7 +244,7 @@ const DepositConfirm: React.FC = () => {
     }
   );
 
-  const depositMutationKey = [...approvedQueryKey, wholeTokenAmount] as const;
+  const depositMutationKey = [...approvedQueryKey, amount] as const;
   const depositMutation = useMutation<BigNumber | undefined, unknown, BigNumber, unknown>(
     depositMutationKey,
     async (unitAmount): Promise<BigNumber | undefined> => {
@@ -328,22 +280,6 @@ const DepositConfirm: React.FC = () => {
     }
   );
 
-  useEffect(() => {
-    if (match.params && match.params.amount) {
-      try {
-        const parsed = Number(String(match.params.amount));
-        if (wholeTokenAmount !== parsed) {
-          setWholeTokenAmount(parsed);
-        }
-      } catch {
-        // Don't set the number if the match path isn't one
-      }
-    }
-    /*
-    if (step != 2 && approval && balance && approval.gte(balance)) {
-      setStep(2);
-    }*/
-  }, [match, approval, balance, wholeTokenAmount, setWholeTokenAmount, step, setStep]);
 
   return (
     <Page>
@@ -366,11 +302,11 @@ const DepositConfirm: React.FC = () => {
                     <div className="token-amount">
                       <img src={asset.img} alt="" />
                       <span>
-                        {wholeTokenAmount} {asset.name}
+                        {amount} {asset.name}
                       </span>
                     </div>
                     <div className="usd-amount">
-                      $ {asset.asset_price * wholeTokenAmount}
+                      $ {asset.asset_price * amount}
                     </div>
                   </div>
                 ) : (
@@ -406,7 +342,7 @@ const DepositConfirm: React.FC = () => {
                     <div className="form-action-body-left">
                       <div className="title">Approve</div>
                       <div className="desc">
-                        Please approve before deposting {wholeTokenAmount} eth
+                        Please approve before depositing {amount} {assetName}
                       </div>
                     </div>
                     <div className="form-action-body-right">
@@ -442,7 +378,7 @@ const DepositConfirm: React.FC = () => {
                         disabled={depositMutation.isLoading || approvalMutation.isLoading || approval === undefined}
                         onClick={() => {
                           depositMutation
-                            .mutateAsync(ethers.utils.parseEther(wholeTokenAmount.toString()))
+                            .mutateAsync(ethers.utils.parseEther(amount.toString()))
                             .then(async (result) => {
                               if (result) {
                                 setStep(step + 1)
